timeLimitsManager: Don't log transitions from history
The history file can get quite large, with the result that we
flood the journal with pages and pages of old transition if
debug messages are enabled.

Instead, suppress the regular debug log and print a summary message
at the end instead.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3740>