runDialog: Set label-actor for entry
The dialog title is currently "invisible" to the screen reader,
so alt+F2 effectively brings up a plain entry with no context.

To address this, set the entry's label-actor to the dialog title,
so we export a proper labelled-by relationship to the screen reader.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3736>