subprojects: Switch to wrap to provide libshew
The Extensions app is very simple, but hard to contribute to
because the default workflow (open in Builder, hit run) does
not work due to the code being in a subtree of gnome-shell.

However it cannot be easily split into its own repository,
while the shew subproject is still provided gnome-shell.

To allow for the future split, libshew was moved into a
separate repo, so remove the in-tree subproject and replace
it with a meson wrap.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3738>