appFavorites: Remove rename list
With the exception of Simple Scan, the last update to this list was 5
years ago. If Simple Scan is one of your favorite apps and you are still
using GNOME 43 or below, then I am sorry. Otherwise, this list is
entirely obsolete.

Don't rename desktop files. It's very disruptive.

If you do need to rename a desktop file for some reason, the user will
just have to pin the favorite app again.