ci: Set variables at job level where possible
Variables that are defined at the global level are more disconnected
from the jobs that use them, and therefore more likely to get outdated
when the job changes.

To litigate that, define variables locally with the job unless they
are used by multiple jobs.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3745>