ci: Run eslint-doc job unconditionally
The job is meant to run when either the style guide or the eslint
configuration changes. That stopped working when the style guide
was renamed to docs/js-coding-style.md without updating the job
conditions, whoops.

The job isn't very expensive to run, so just run it unconditionally
to avoid these kinds of oversights in the future.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3745>