brightnessManager: Add keyboard shortcuts for monitor brightness control
This used to be done by g-s-d but to avoid the split brain situation,
g-s-d has to stop poking at the backlight directly. This also means that
there is no reason to round-trip through g-s-d anymore, and we can just
implement the shortcuts in gnome shell.

This also adds new shortcuts for controling the backlight of the "active
monitor" (i.e. the one with the cursor) by pressing shift as well.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3751>