authPrompt: Notify error messages in GDM login/unlock
When there is an error during GDM login or unlock (for example,
typing a wrong password), a text message is shown, but nothing
is said by the screen reader.

This patch fixes this, by notifying any error set in the auth
prompt.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3765>