st/entry: Forward label-actor to internal ClutterText
The StWidget:label-actor property is a convenient way of setting up an
actor's labelled-by relationship. However on StEntry, it is currently
fairly useless because if the entry is focused, the actual keyboard
focus will be on the internal ClutterText rather than the entry itself,
and orca ignores ancestors in most cases.

Address this by tracking the entry's label-actor property, and replicate
the labelled-by relationship on the internal ClutterText actor.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3179>