loginDialog: Increase size of session/a11y buttons
The buttons on the login screen for the session and accessibility
menus can easily be lost, due to a combination of their size and
position, and the large amount of empty space.

Increase their size a bit to lessen that effect.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8497
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3794>