switcherPopup: Move to a ClutterGesture for closing the popup
When the user presses a mouse button outside of the switcherPopup, the
popup should close. This broke when we moved StButton to ClutterGesture,
because the button_press_event() handler in switcherPopup will close the
popup before the click gesture recognizes.

Since we're trying to move away from direct event handling anyway, port
this functionality to a LongPressGesture with a duration of 0, so that
the popup closes immediately on button press as it used to.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2853>