portalHelper: Schedule recheck if it's too soon to recheck
Currently we have a 30 second limit on rechecks. A recheck that would
violate the limit is simply dropped. This causes the portal helper to
never close itself.

Instead, schedule a future recheck for as soon as the timeout expires.
This will put an upper bound on how long the portal helper will
unnecessarily remain open.

(I've also changed the logic to unconditionally emit CANCELLED in
vfunc_close_request(), since this seems more correct. This won't
actually affect anything, though, because gnome-shell just ignores the
CANCELLED status.)

Closes https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7790