swipeTracker: Port to use ClutterPanGesture
The new ClutterGesture subclass ClutterPanGesture fits the SwipeTracker
quite well. API wise the biggest difference to the TouchSwipeGesture it
was using before is that ClutterPanGesture has its own internal event
history to keep track of deltas and to calculate a velocity in the end.
That means we need to give it slighty special treatment right now, but
we'll improve that with the next commits.

So replace the internal TouchSwipeGesture and make the SwipeTracker use
ClutterPanGesture directly.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2853>