Set event phase to CAPTURE for swipe gestures starting in the session
When a swipe gesture can get triggered in the session (ie. on top of a
window, not just from the overview or other shell chrome), we want it to
receive events as soon as possible, as otherwise the clients will see
those events.

So use the CAPTURE phase for the overview gesture and the workspace
switching gesture, which makes sure the gesture sees each event before a
window gets to see it.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2853>