brightnessManager: Fix stepping up/down via shortcuts
The `BrightnessScale.value` property was originally called `brightness`,
and the methods used for keyboard shortcuts somehow missed the name change.

As a result, the methods try to set the brightness to `undefined`, which
is out of range and thus fails.

Use the correct property name to fix the shortcuts.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8570
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3825>