status/network: Update Wi-Fi quicksettings icon on disconnect
Prior to this patch, upon a connected AP disconnecting from
the device, the current quicksettings wi-fi icon will continue
showing the "network-wireless-connected-symbolic" symbol in
the time between the initial disconnection and the adapter
deactivation, even though the connection has been
degraded by virtue of no longer being routable.

This will also happen in the
NM.ActiveConnectionState.DEACTIVATING state.

To make this clearer to the user, this change modifies
quicksettings' wifi icon to reflect the degraded states:
- An activated adapter with no active AP connection will
now map to the network-wireless-offline-symbolic icon
(to indicate that the AP signal is lost)

- NM.ActiveConnectionState.DEACTIVATING will now map to the
network-wireless-acquiring-symbolic icon (to indicate that
a state transition is in progress)

Closes https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8605

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3840>