parentalControlsManager: Add support for session limits
Add ability to check if session limits are enabled for users using API
introduced in malcontent 0.14.0. The `get_session_limits_async` function
is actually available since 0.5.0, but the signal
`session-limits-changed` was only introduced in 0.14.0, so check for
function `get_daily_limit` which was introduced in 0.14.0 to be sure
we're using correct version. This allows us to keep filtering apps, even
if the library version doesn't fully support session limits.

Will be used in the future to integrate parental controls session
limits.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3841>