authPrompt: Fix key focus handling on choice list
When a choice list widget is prompted in GDM we call updateSensitivity()
but this does not act on the currently visible authentication widget but
rather on the text entry all the times.

And this causes that when a choice list is shown we always set the key
focus to the (hidden) text entry.

This implies that when the list is finally shown, moving the arrow keys
leads makes the entry context menu to be shown instead of being able to
navigate through the entries.

So, make updateSensitivity use the currently visible authorization
widget instead and use it to control the sensitivity of the choice list