extensionSystem: Fix race condition when enabling extensions
When an extension is enabled, it is deleted from the list of
disabled extensions and added to the list of enabled extensions.
This can cause a race condition, because both cases call the
same asynchronous function, resulting in the creation of two
instances of each extension and the `enable` method of each
one being called twice. This is possible because the method
`onEnabledExtensionsChanged` can be called both from a signal
of a change in the `disabled-extensions` property, or from
a signal of a change in the `enabled-extensions` property.
Since this method is async and does several `await` calls,
two very close changes to those properties can trigger it
the second time when the first time is in the middle of an
`await`, and, thus, in an incomplete state.

Fix this by using the previously defined `asyncMutex` object.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3640>