tests/unit: Add a helper method in TimeLimitsManager
Currently we create a new `TimeLimitsManager` for each of the unit tests
for testing the wellbeing screen time limit, with the same parameters.
Rather than having to repeat the entire constructor call multiple times,
factor it out into a helper method on the `TestHarness`. There is no
difference in functionality itself.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3841>