docs: Replace references to --nested with --devkit
The --devkit option is strongly recommended over the old --nested
nowadays, not least because the latter option no longer exists by
default (as it is tied to the X11 backend).

Remove the section on global shortcuts and workaround for looking glass,
as Devkit supports inhibiting keyboard shortcuts.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3871>