gdm: Add support for Passkey in authServicesSSSDSwitchable
This allows selecting passkey authentication mechanism.

Use passkeyManager to detect when there's a passkey inserted or removed
to restart the service so the JSON protocol gives valid information
about the passkey.

It's used 'show-choice-list' with an emtpy list, to display the touch
instruction.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3212>