gdm: Add support for Passkey in authServicesSwitchable
This allows selecting passkey authentication mechanism.

Currently it will only be displayed for the user that has a passkey mapped and
when the key is plugged in.
It is needed to consider when the key is not plugged in to display the
message "Insert security key", and allow selecting even when the key is not
plugged in.

It's leveraged using 'show-choice-list' with an emtpy list, to display the touch
instruction.