gdm/util: Allow null _hold and don't recreate dummy holds
_hold property is used to inform the caller of begin method (authPrompt)
when it's started the verification (calling acquire) and when it
finished starting it, either successfully or with an error (calling
release).

There's no need to create dummy holds that won't be used anywhere
externally.

In the next commits the use of _hold will be simplified considerably,
but for now, just accept getting null and dont create dummy ones.