gdm: Add PasskeyDeviceManager
This utility will be used in the next commits, when passkey authentication
is implemented, to detect when a passkey has been inserted or removed.

To detect if a sysfs device is a passkey (fido2), it's been used the
implementation of systemd in fido_id_desc.c.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3212>