authPrompt: Update authList style
Make the buttons a bit bigger and more rounded.

Buttons can have two labels and one icon.
The icon is used to inform about Organization field of the certificate when
hovering over it. This will mainly be used by the next commit, for the
moment keep relying only in one single label (update choiceList list
format).

Now the title is an insensitive button that is in _mainBox to ensure
back button is visible and properly aligned.

Add accessible_name.