messageList: Make media messages follow notification policy
Unlike regular notifications, media messages are shown unconditionally
for any active MPRIS player. In the past that was uncontroversial, but
now that media messages are shown on the lock screen as well, there is
a privacy concern.

To address this, make media messages follow the notification policy
of the corresponding app (if specified) or the general notification
settings.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8600
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3896>