i18n: Add context to "Power mode" title string
The "Power mode" string in the quick settings menu appears both in the
title of the button and as the menu header that opens when clicking on
such button.
Currently they are translated together, but---at least in
Italian---the standard translation, "Modalità energia", is too long to
fit in the button and is very often truncated to "Modalità…" (aka
"Mode", in English) from which is not clear what the menu is about.
We would use "Energia" for the button title, but at the same time it
would be nice to have the full "Modalità energia" translation in the
menu header, where there is more space available.

For this reason I propose to add contexts to the two "Power mode"
strings, so that the string can be translated in different ways
depending on its position.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3903>