grabHelper: Port to using ClickGesture to dismiss
There's a few things we don't need anymore here. We don't
need holding on the captured events until button release
anymore, and we don't need to track crossing events outside
the grab as those will not get emitted as usual. This allows
for a few simplifications in handling the events for the
GrabHelper.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3912>