screenshot: Make screenshot tooltips label actors for their widgets
This ensures that the buttons have a label from their tooltip if we do not
have another label actor already (like from the icon label buttons).

Because no button using a toolltip in this UI has a label on its own,
this is not checked.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3908>