status/network: Update indicator icon on connectivity changes
When there is no active main connection or it uses a device type we do
not handle, the indicator icon and its visibility depend on the global
connectivity.

However we currently don't update the icon on connectivity changes,
with the result that we sometimes show the fallback connection icon
when the network is in fact disconnected.

To fix this, update the icon whenever the (global) connectivity changes.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7357
Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8549
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3910>
(cherry picked from commit a999193c06ad91a40d9eae5f01c2ccaf9455c227)