brightnessManager: Use slider as bias when auto brightness is being used
The code previously tried to ensure that the sliders were disabled when
auto brightness was active, and determine the brightness just from the
auto brightness target we get from g-s-d.

It turns out that users do want to still have some control over the
brightness, even if auto brightness is active.

The slider is now being used as a bias for the auto brightness target
instead.

This also means that we can drop the BrightnessScale::locked property.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3935>